Refactor Overlay for Structured Network support: * Move overlay up one directory * Add abstract_protocol_handler, message_stream * Add peer_protocol_detector * Tidy up some declarations * Use strand::running_in_this_thread instead of bool * Update README.md * Replace protocol message read loop:   - Process data in arbitrary size chunks   - message_stream extracts individual messages   - peer_protocol_detector identifies the handshake   - abstract_protocol_handler used for dispatching messages * Remove unused protocol message types:   - mtACCOUNT   - mtCONTACT   - mtERROR   - mtGET_ACCOUNT   - mtGET_CONTACTS   - mtGET_VALIDATIONS   - mtSEARCH_TRANSACTION   - mtUNUSED_FIELD
Conflicts:
	src/ripple/module/app/main/Application.cpp
	src/ripple/module/app/misc/NetworkOPs.cpp
	src/ripple/module/app/peers/PeerSet.cpp
