+++ src/ripple/module/app/ledger/LedgerMaster.cpp
@@ -550,7 +550,7 @@ public:
             tmBH.set_ledgerhash (nextLedger->getHash().begin (), 32);
             Message::pointer packet = std::make_shared<Message> (tmBH, protocol::mtGET_OBJECTS);
 
-            target->sendPacket (packet, false);
+            target->send (packet);
             WriteLog (lsTRACE, LedgerMaster) << "Requested fetch pack for " << nextLedger->getLedgerSeq() - 1;
         }
         else
