+++ src/ripple/proto/ripple.proto
@@ -2,40 +2,30 @@ package protocol;
 
 enum MessageType
 {
-    // core
     mtHELLO                 = 1;
-    mtERROR_MSG             = 2;
     mtPING                  = 3;
     mtPROOFOFWORK           = 4;
     mtCLUSTER               = 5;
-
-    // network presence detection
-    mtGET_CONTACTS          = 10;
-    mtCONTACT               = 11;
     mtGET_PEERS             = 12;
     mtPEERS                 = 13;
-    // This field was previous used for PeerFinder ENDPOINTS messages. Since the
-    // structure's required fields changed, the message number was changed.
-    mtUNUSED_FIELD          = 14;
     mtENDPOINTS             = 15;
-
-    // operations for 'small' nodes
-    mtSEARCH_TRANSACTION    = 20;
-    mtGET_ACCOUNT           = 21;
-    mtACCOUNT               = 22;
-
-    // transaction and ledger processing
     mtTRANSACTION           = 30;
     mtGET_LEDGER            = 31;
     mtLEDGER_DATA           = 32;
     mtPROPOSE_LEDGER        = 33;
     mtSTATUS_CHANGE         = 34;
     mtHAVE_SET              = 35;
-
-    // data replication and synchronization
-    mtGET_VALIDATIONS       = 40;
     mtVALIDATION            = 41;
     mtGET_OBJECTS           = 42;
+
+    // <available>          = 2;
+    // <available>          = 10;
+    // <available>          = 11;
+    // <available>          = 14;
+    // <available>          = 20;
+    // <available>          = 21;
+    // <available>          = 22;
+    // <available>          = 40;
 }
 
 // token, iterations, target, challenge = issue demand for proof of work
@@ -199,31 +189,6 @@ message TMValidation
     optional bool checkedSignature  = 2;        // node vouches signature is correct
 }
 
-message TMGetValidations
-{
-    required uint32 ledgerIndex     = 1;
-    repeated bytes hanko            = 2;
-    optional uint32 count           = 3; // get random validations
-}
-
-message TMContact
-{
-    required bytes pubKey               = 1;
-    required uint32 softwareVersion     = 2;
-    required uint32 protoVersion        = 3;
-    required uint64 nodeFlags           = 4;
-    required uint64 timestamp           = 5;
-    repeated bytes nodeInfo             = 6;
-    required bytes signature            = 7;
-}
-
-// request node information
-message TMGetContacts
-{
-    repeated bytes nodeIDs          = 1; // specific nodes we want
-    optional uint32 nodeCount       = 2; // get some random nodes
-}
-
 message TMGetPeers
 {
     required uint32 doWeNeedThis    = 1;  // yes since you are asserting that the packet size isn't 0 in Message
@@ -260,36 +225,6 @@ message TMEndpoints
 
     repeated TMEndpoint     endpoints = 2;
 };
-
-message TMSearchTransaction
-{
-    required uint32 maxTrans    = 1;
-    optional bytes toAccount    = 2;
-    optional bytes fromAccount  = 3;
-    optional uint32 minLedger   = 4;
-    optional bytes fromAcctSeq  = 5;
-    repeated bytes transID      = 6;
-}
-
-message TMGetAccount
-{
-    repeated bytes acctID       = 1;
-    optional uint32 seq         = 2;
-}
-
-message Account
-{
-    required bytes accountID    = 1;
-    required uint64 balance     = 2;
-    required uint32 accountSeq  = 3;
-    required uint32 ledgerSeq   = 4;
-}
-
-message TMAccount
-{
-    repeated Account accounts   = 1;
-    optional uint32 seq         = 2;
-}
 
 message TMIndexedObject
 {
@@ -386,9 +321,3 @@ message TMPing
     optional uint64 netTime     = 4;
 }
 
-
-message TMErrorMsg
-{
-    optional int32 errorCode    = 1;
-    optional string message     = 2;
-}
